feat: Add noise values on payload data
The XDK Workbench v.3.6.1 enables the use of noise sensor alongside of other sensors.
The current version of noise on MITA only returns the noise value on terms of milli volts.

TODO: Look for a function on MITA to convert milli volts data to decibels, watch out, it appears complex math operations are not enabled on MITA (log, exp, and so on)